pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langdeep"
version = "0.1.3"
description = "LangChain + LangGraph Agent Workflow Framework"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ZChunzi", email = "rzhou990@gmail.com"}
]
keywords = ["langchain", "langgraph", "agent", "orchestration", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "pyyaml>=6.0",
    "croniter>=2.0.0",
]

[project.optional-dependencies]
all = [
    "langchain-anthropic",
    "langchain-google-genai",
    "langchain-ollama",
    "langchain-tavily",
]
persist = ["SQLAlchemy>=2.0"]

[project.urls]
Repository = "https://github.com/ZChunzi/langdeep"

[tool.setuptools]
package-dir = {"langdeep" = "src"}
packages = [
    "langdeep",
    "langdeep.core",
    "langdeep.core.decorators",
    "langdeep.core.registry",
    "langdeep.core.orchestrator",
    "langdeep.core.planner",
    "langdeep.core.prompt",
    "langdeep.core.scheduling",
    "langdeep.schemas",
    "langdeep.utils",
]
include-package-data = true

[tool.setuptools.package-data]
"langdeep" = ["resources/prompts/*.md"]